+++ src/Context.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+export const QuestionsContext = createContext([]);
+export const InterviewsContext = createContext([]);
+export const UserContext = createContext('');
