+++ src/components/StudentQuestionContainer/StudentQuestionContainer.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const StudentQuestionContainer = () => {
+  return (
+    <div>Hello</div>
+  )
+}
+
+export default StudentQuestionContainer;
